cli: Change default program behavior to iterate over all projects
arch_release_promotion/cli.py:
Add `promote_project_release()` to promote releases of projects
generically. If no `release_version` is provided, user input is
requested to select a release that is not yet promoted.
Change `main()` to rely on `promote_project_release()` by either
promoting a specific release of a project (if the project or the project
and the release version are provided) or iterating over all available
projects in the configuration to allow the user to promote releases for
them.