Update functionally blocking dependencies
pyproject.toml:
Update the python requirement to 3.10.
Update the prometheus-client requirement to 0.14.1.
Update the pytest dev requirement to 7.1.
Update the black dev requirement to 22.3.
Update the mypy dev requirement to 0.942.
Update the flake8 dev requirement to 4.0.1.
Update the coverage dev requirement to 6.1.
Change the build-system to also pull in setuptools as some dependencies
do not build properly.
Change the mypy configuration to show specific error codes.